refactor: move Effect and FnEffect imports to base module and update references
