+++ src/pyeffects/base.py
